7단계 미션 완료
- auth 패키지를 roomescape 외부로 분리
- AuthController의 로직을 MemberController에 병합
- @ComponentScan이 붙은 @SpringBootApplication이 위치한 roomescape 패키지 외부에 있는 설정 파일도 스캔하기 위해 @Import(설정파일 추가에 사용)를 사용
