+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -1,8 +1,11 @@
 package roomescape;
 
+import auth.AuthConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
 
+@Import(AuthConfig.class)
 @SpringBootApplication
 public class RoomescapeApplication {
     public static void main(String[] args) {
